+++ src/content.tsx
@@ -1,3 +1,4 @@
+import { getPreferenceValues } from '@raycast/api';
 import { ActionPanel, Action, List, Icon } from '@raycast/api';
 import { useEffect, useState, useMemo } from 'react';
 import { decrypt } from './pass';
@@ -32,6 +33,7 @@ function parseRows(content: string): Row[] {
 }
 
 function ContentRow(row: Row) {
+  const { defaultAction } = getPreferenceValues();
   const [show, setShow] = useState<boolean>(false);
 
   const { toggleTitle, toggleIcon, itemTitle } = useMemo(() => {
@@ -55,8 +57,17 @@ function ContentRow(row: Row) {
       title={itemTitle}
       actions={
         <ActionPanel>
-          <Action.Paste content={row.value} />
-          <Action.CopyToClipboard content={row.value} shortcut={{ modifiers: ['cmd', 'shift'], key: 'c' }} />
+          {defaultAction === 'copy' ? (
+            <>
+              <Action.CopyToClipboard content={row.value} />
+              <Action.Paste content={row.value} />
+            </>
+          ) : (
+            <>
+              <Action.Paste content={row.value} />
+              <Action.CopyToClipboard content={row.value} />
+            </>
+          )}
           <Action
             icon={toggleIcon}
             title={toggleTitle}